+++ src/processor/mod.rs
@@ -86,7 +86,7 @@ pub trait EliminateIF: Instantiate {
         C: ClauseDBIF,
         R: RestartIF;
     /// return the order of vars based on their occurrences
-    fn sorted_iterator(&self) -> Iter<'_, usize>;
+    fn sorted_iterator(&self) -> Iter<'_, u32>;
     /// return vi's stats
     fn stats(&self, vi: VarId) -> Option<(usize, usize)>;
     /// return the constraints on eliminated literals.
